test(syncer): split RDBMS tests into MySQL and PostgreSQL files
